+++ packages/mattermost-redux/src/actions/cloud.ts
@@ -4,11 +4,8 @@
 import {CloudTypes} from 'mattermost-redux/action_types';
 import {Client4} from 'mattermost-redux/client';
 
-import {DispatchFunc, GetStateFunc, ActionFunc} from 'mattermost-redux/types/actions';
+import {ActionFunc} from 'mattermost-redux/types/actions';
 import {Address, CloudCustomerPatch} from '@mattermost/types/cloud';
-
-import {getCloudErrors} from 'mattermost-redux/selectors/entities/cloud';
-import {getCloudLimits} from 'actions/cloud';
 
 import {bindClientFunc} from './helpers';
 
@@ -47,37 +44,6 @@ export function getInvoices(): ActionFunc {
         onFailure: CloudTypes.CLOUD_INVOICES_FAILED,
         onRequest: CloudTypes.CLOUD_INVOICES_REQUEST,
     });
-}
-
-export function retryFailedCloudFetches() {
-    return (dispatch: DispatchFunc, getState: GetStateFunc) => {
-        const errors = getCloudErrors(getState());
-        if (Object.keys(errors).length === 0) {
-            return {data: true};
-        }
-
-        if (errors.subscription) {
-            dispatch(getCloudSubscription());
-        }
-
-        if (errors.products) {
-            dispatch(getCloudProducts());
-        }
-
-        if (errors.customer) {
-            dispatch(getCloudCustomer());
-        }
-
-        if (errors.invoices) {
-            dispatch(getInvoices());
-        }
-
-        if (errors.limits) {
-            getCloudLimits()(dispatch, getState);
-        }
-
-        return {data: true};
-    };
 }
 
 export function updateCloudCustomer(customerPatch: CloudCustomerPatch): ActionFunc {
